C++FantasyGame
Your task is to write a simple Fantasy Game played through a text
interface. The aim of the game is to collect as much gold as possible
by defeating enemy characters. A (partially) implemented executable can
be found on the module’s SULIS page to give you an idea of what is
expected. If you have any queries, please post them to the Questions &
Answers Forum.